% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCountsPred.R
\name{calculateCountsPred}
\alias{calculateCountsPred}
\title{Extract first order moments values from ODEs solutions}
\usage{
calculateCountsPred(sdePred, ncell)
}
\arguments{
\item{sdePred}{Differential equations solutions.}

\item{ncell}{The number of cell types over which the process evolves.}
}
\value{
Vector of predicted cell counts.
}
\description{
Extract first order moments values from ODEs solutions
}
\examples{
calculateCountsPred(sdePred,ncell)
}
