% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJ.R
\name{calculateJ}
\alias{calculateJ}
\title{Calculate the Jacobian matrix}
\usage{
calculateJ(initialValues1_2ODEs, a, ncell, eps)
}
\arguments{
\item{initialValues1_2ODEs}{Differential equations initial values.}

\item{a}{ODEs parameters values.}

\item{ncell}{The number of cell types over which the process evolves.}

\item{eps}{Delta for numeric derivatives calculation}
}
\value{
Jacobian matrix.
}
\description{
Calculate the Jacobian matrix
}
\examples{
calculateJ(initialValues1_2ODEs,a,ncell,eps)
}
