% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePred.R
\name{calculatePred}
\alias{calculatePred}
\title{Calculate the solution of the ODEs}
\usage{
calculatePred(initialValues1_2ODEs, a)
}
\arguments{
\item{initialValues1_2ODEs}{Differential equations initial values.}

\item{a}{ODEs parameters values.}
}
\value{
List of prediction.
}
\description{
Solve the system of differential equations given initial values and parameters.
}
\examples{
calculatePred(initialValues1_2ODEs,a)
}
