% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateW1Inv.R
\name{calculateW1Inv}
\alias{calculateW1Inv}
\title{Calculate the global block diagonale covariance matrix inverse from ODEs solutions}
\usage{
calculateW1Inv(sdePred, lambda = 0, ncell, combOk)
}
\arguments{
\item{sdePred}{Differential equations solutions.}

\item{lambda}{Additional variance to be added on the main diagonal of the global covariance matrix.}

\item{ncell}{The number of cell types over which the process evolves.}

\item{combOk}{The number of cell types over which the process evolves.}
}
\value{
Block diagonal matrix.
}
\description{
Combines differential equation solutions and calculate the inverse of covariance matrix W. It is possible to add additional variance on the main diagonal (addtional source of noise) by setting lambda.
}
\examples{
calculateW1Inv(sdePred,lambda=0,ncell,combOk)
}
