% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcW.R
\name{calcW}
\alias{calcW}
\title{Calculate covariance matrix}
\usage{
calcW(diffEqSol, ncell, combOk)
}
\arguments{
\item{diffEqSol}{Differential equations solutions.#'}

\item{ncell}{The number of cell types over which the process evolves.}

\item{combOk}{The number of cell types over which the process evolves.}
}
\value{
Matrix.
}
\description{
Combines differential equation solutions as encoded in combOk matrix and calculate the covariance matrix W
}
\examples{
calcW(diffEqSol,ncell,combOk)
}
