% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dXCalc.R
\name{dXCalc}
\alias{dXCalc}
\title{Calculate the differences in cell counts among consecutive timepoints.}
\usage{
dXCalc(clonesTrajs)
}
\arguments{
\item{clonesTrajs}{Set of clone trajectories.}
}
\value{
Deltas matrix.
}
\description{
Calculate the differences in cell counts among consecutive observations for a set of clone trajectories. The output is in the OLS procedure to set parameters initial values.
}
\examples{
dXCalc(clonesTrajs)
}
