% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endValuesColVec.R
\name{endValuesColVec}
\alias{endValuesColVec}
\title{Extract clones observations at time points t= 1 .. n.}
\usage{
endValuesColVec(clonesTrajs)
}
\arguments{
\item{clonesTrajs}{Set of clone trajectories.}
}
\value{
Matrix.
}
\description{
Extract clones observations at time points t= 1 .. n. These are the values to be matched by ODEs solutions.
}
\examples{
endValuesColVec(clonesTrajs)
}
