% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOdeFormula.R
\name{generateOdeFormula}
\alias{generateOdeFormula}
\title{Generate SLCDP ODE formulas for moments evolution.}
\usage{
generateOdeFormula(ncell)
}
\arguments{
\item{ncell}{The number of cell types over which the process evolves.}
}
\value{
Text corresponding to formula.
}
\description{
Automatically generate the ODE formulas describing the first and second-order moments evolution for a SLCDP with ncell cell types.
}
\examples{
generateOdeFormula(5)
}
