% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonesTrajsStartValues.R
\name{initialValues1_2ODEsCalc}
\alias{initialValues1_2ODEsCalc}
\title{Calculate the starting values for moments ODEs.}
\usage{
initialValues1_2ODEsCalc(clonesTrajsStartValues, ncell)
}
\arguments{
\item{clonesTrajsStartValues}{Clone trajectories starting value returned by startValues()}

\item{ncell}{The number of cell types over which the process evolves.}
}
\description{
Calculate the starting values for moments ODEs.
}
\examples{
initialValues1_2ODEsCalc(clonesTrajsStartValues,ncell)
}
