% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mXCalc.R
\name{mXCalc}
\alias{mXCalc}
\title{Calculate the linear predictor matrix M for the local linear approximation.}
\usage{
mXCalc(clonesTrajs, netMat, dt, ncell)
}
\arguments{
\item{clonesTrajs}{Set of clone trajectories.}

\item{netMat}{Net effect matrix.}

\item{dt}{Time between observation}

\item{ncell}{The number of cell types over which the process evolves.}
}
\value{
Predictors matrix.
}
\description{
By combining cell count observations, net effect matrix and dt it returns the matrix of linear predictors for setting parameters initial values.
}
\examples{
mXCalc(clonesTrajs,netMat,dt,ncell)
}
