% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netEffectMatrixCalc.R
\name{netEffectMatrixCalc}
\alias{netEffectMatrixCalc}
\title{Generate the net effect matrix for the SLCDP.}
\usage{
netEffectMatrixCalc(ncell)
}
\arguments{
\item{ncell}{The number of cell types over which the process evolves.}
}
\value{
r x ncell matrix.
}
\description{
Generate the r x ncell net effect matrix describing the change in cell counts induces by each of the r possible cellular events. The first ncell rows correspond to duplication events (+1 in process state), followed by ncell death events (-1) and all possible differentiation (-1;+1 respectively for the cell type of origin and destination). (). , the  ODE formulas describing the first and second-order moments evolution for a SLCDP with ncell cell types.
}
\examples{
netEffectMatrixCalc(5)
}
