% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOptimizationOptiSolve.R
\name{runOptimizationOptiSolve}
\alias{runOptimizationOptiSolve}
\title{Calculate the solution for the Quadratic Programming (QP) problem using Cplex optimization environment.}
\usage{
runOptimizationOptiSolve(MX, dX, W1Inv = NULL, constraints, bvec = NULL)
}
\arguments{
\item{MX}{Predictor matrix. It can be M matrix in OLS or Jacobian matrix in the Gauss-Newton algorithm.}

\item{dX}{Response vector. It can be deltas vector OLS or endValues in the Gauss-Newton algorithm.}

\item{W1Inv}{Inverse of the estimated covariance matrix. If NULL it is set to the appropriate identity matrix.}

\item{constraints}{list containing the constraints as generated by setConstraints.}

\item{bvec=NULL}{values for the inequality constraints (rates >= bvec). If NULL, bvec is filled with 0s.}
}
\value{
QP solution.
}
\description{
It combines input matrices and constraints for solving the QP problem.
}
\examples{
runOptimizationOptiSolve(MX,dX,W1Inv=NULL,constraints,bvec=NULL)
}
