% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secOrdMoments.R
\name{secOrdMoments}
\alias{secOrdMoments}
\title{Combination for second order moments}
\usage{
secOrdMoments(ncell)
}
\arguments{
\item{ncell}{The number of cell types over which the process evolves.}
}
\value{
Matrix.
}
\description{
Combination for second order moments
}
\examples{
secOrdMoments(ncell)
}
