% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConstraintsCplex.R
\name{setConstraintsCplex}
\alias{setConstraintsCplex}
\title{Create a list containing rates constraints and bounds for the definition of the Quadratic Programming problem in the Cplex optimization environment.}
\usage{
setConstraintsCplex(
  ncell,
  duplRatesLB = NA,
  duplRatesUB = NA,
  deathRatesLB = NA,
  deathRatesUB = NA,
  diffRatesLB = NA,
  diffRatesUB = NA
)
}
\arguments{
\item{ncell}{The number of cell types over which the process evolves.}

\item{duplRatesLB}{duplRatesUB deathRatesLB deathRatesUB Vectors of length ncell containing duplication (dupl) and death (death) rates upper- (UB) and lower- (LB) bounds.}

\item{diffRatesLB}{diffRatesUB Matrix of size ncell x ncell containing differentiation rates upper- (UB) and lower- (LB) bounds.}
}
\value{
List with A, (inequality constraints),lb (lower bound), ub (upper bound).
}
\description{
It initializes and stores constraints and bounds for each individual rates. By default, given ncell, it sets all reates to be non-negative, exept for differentiation rates matrix main diagonal elements set to 0. Using this function it is possible to set specific rates to fixed values (non-negative).
}
\examples{
setConstraintsCplex(ncell,duplRatesLB=NA,duplRatesUB=NA,deathRatesLB=NA,deathRatesUB=NA,diffRatesLB=NA,diffRatesUB=NA)
}
