% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startValues.R
\name{startValues}
\alias{startValues}
\title{Extract clones observations at time points t= 0 .. (n-1).}
\usage{
startValues(clonesTrajs)
}
\arguments{
\item{clonesTrajs}{Set of clone trajectories.}
}
\value{
Matrix.
}
\description{
Extract clones observations at time points t= 0 .. (n-1), useful for the calculation of ODEs starting values.
}
\examples{
startValues(clonesTrajs)
}
